# src/static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TaskFlow API",
    "description": "A secure and scalable task management API for managing tasks and categories.",
    "contact": {
      "email": "support@taskflow.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "2.0.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
...
  "paths": {
...
    "/tasks": {
      "get": {
        "tags": ["Tasks"],
        "summary": "Get all tasks (Web UI)",
        "operationId": "getTasks",
        "responses": {
          "200": {
            "description": "List of tasks displayed successfully"
          },
          "401": {
            "description": "Authentication required"
          }
        }
      }
    },
    "/tasks/dashboard": {
      "get": {
        "tags": ["Tasks"],
        "summary": "Display task dashboard (Web UI)",
        "operationId": "getDashboard",
        "responses": {
          "200": {
            "description": "Dashboard displayed successfully"
          },
          "401": {
            "description": "Authentication required"
          }
        }
      }
    },
    "/tasks/add_task": {
      "get": {
        "tags": ["Tasks"],
        "summary": "Display task creation form",
        "operationId": "getAddTaskForm",
        "responses": {
          "200": {
            "description": "Task form displayed successfully"
          },
          "401": {
            "description": "Authentication required"
          }
        }
      },
      "post": {
        "tags": ["Tasks"],
        "summary": "Add a new task (Web UI)",
        "operationId": "addTask",
        "requestBody": {
          "description": "Create a new task",
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Task"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Task added successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          }
        }
      }
    },
    "/tasks/edit_task/{taskId}": {
      "get": {
        "tags": ["Tasks"],
        "summary": "Display task edit form",
        "operationId": "getEditTaskForm",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Edit form displayed successfully"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Task not found"
          }
        }
      },
      "post": {
        "tags": ["Tasks"],
        "summary": "Edit an existing task (Web UI)",
        "operationId": "editTask",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "Edit an existing task",
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Task"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Task updated successfully"
          },
          "400": {
            "description": "Error editing task"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Task not found"
          }
        }
      }
    },
    "/tasks/delete_task/{taskId}": {
      "post": {
        "tags": ["Tasks"],
        "summary": "Delete an existing task (Web UI)",
        "operationId": "deleteTask",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "description": "Unique Task Id to delete",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task deleted successfully"
          },
          "400": {
            "description": "Error deleting task"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Task not found"
          }
        }
      }
    },
    "/api/tasks": {
      "get": {
        "tags": ["API Tasks"],
        "summary": "Get all tasks",
        "operationId": "apiGetTasks",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by task status",
            "schema": {
              "type": "string",
              "enum": ["pending", "in-progress", "completed"]
            }
          },
          {
            "name": "category_id",
            "in": "query",
            "description": "Filter by category ID",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "priority",
            "in": "query",
            "description": "Filter by priority",
            "schema": {
              "type": "string",
              "enum": ["low", "medium", "high"]
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Items per page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of tasks",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["API Tasks"],
        "summary": "Create a new task",
        "operationId": "apiCreateTask",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Task details",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Task created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{taskId}": {
      "get": {
        "tags": ["API Tasks"],
        "summary": "Get task by ID",
        "operationId": "apiGetTask",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "description": "Task ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["API Tasks"],
        "summary": "Update task",
        "operationId": "apiUpdateTask",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "description": "Task ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "Updated task details",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Task updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["API Tasks"],
        "summary": "Delete task",
        "operationId": "apiDeleteTask",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "description": "Task ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Task": {
        "type": "object",
        "required": ["title", "description", "due_date", "priority"],
        "properties": {
          "title": {
            "type": "string",
            "example": "Complete project proposal"
          },
          "description": {
            "type": "string",
            "example": "Write a detailed proposal for the new client project"
          },
          "due_date": {
            "type": "string",
            "format": "date-time",
            "example": "2025-06-15T14:00:00Z"
          },
          "priority": {
            "type": "string",
            "enum": ["low", "medium", "high"],
            "example": "high"
          },
          "category_id": {
            "type": "integer",
            "example": 1
          }
        }
      },
...
      "TaskRequest": {
        "type": "object",
        "required": ["title", "description"],
        "properties": {
          "title": {
            "type": "string",
            "example": "Complete project proposal"
          },
          "description": {
            "type": "string",
            "example": "Write a detailed proposal for the new client project"
          },
          "status": {
            "type": "string",
            "enum": ["pending", "in-progress", "completed"],
            "default": "pending",
            "example": "pending"
          },
          "priority": {
            "type": "string",
            "enum": ["low", "medium", "high"],
            "default": "medium",
            "example": "high"
          },
          "due_date": {
            "type": "string",
            "format": "date-time",
            "example": "2025-06-15T14:00:00Z"
          },
          "category_id": {
            "type": "integer",
            "example": 1
          }
        }
      },
...
      "TaskListResponse": {
        "type": "object",
        "properties": {
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskResponse"
            }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "total": {
                "type": "integer",
                "example": 42
              },
              "page": {
                "type": "integer",
                "example": 1
              },
              "per_page": {
                "type": "integer",
                "example": 20
              },
              "pages": {
                "type": "integer",
                "example": 3
              },
              "has_next": {
                "type": "boolean",
                "example": true
              },
              "has_prev": {
                "type": "boolean",
                "example": false
              }
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "An error occurred"
          },
          "details": {
            "type": "string",
            "example": "Detailed error description"
          }
        }
      },
...
    }
  }
}